cron/requirements.txt
selenium==3.8.0
urllib3==1.22
requests==2.18.4
PyJWT==1.7.1